refactor(store-service): refactoring user-id hardcode
store-service에 user-id 하드코딩 된 코드를 access token에 담긴 userid로 변경
