--- owner-vue/src/util/customUtil.js
@@ -1,9 +1,0 @@
-export default {
-    deepCopy : function (o) {
-        let result = {};
-        if (typeof o === "object" && o !== null)
-            for (let i in o) result[i] = this.deepCopy(o[i]);
-        else result = o;
-        return result;
-    }
-}
