+++ store-service/src/main/java/com/justpickup/storeservice/domain/category/entity/Category.java
@@ -30,7 +30,7 @@ public class Category extends BaseEntity {
     @JoinColumn(name = "store_id")
     private Store store;
 
-    @OneToMany(mappedBy = "category")
+    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
     private List<Item> items;
 
     // == 연관관계 편의 메소드 == //
