+++ user-service/src/main/java/com/justpickup/userservice/UserServiceApplication.java
@@ -8,6 +8,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.context.annotation.Bean;
 
+import java.util.List;
+
 @SpringBootApplication
 @EnableEurekaClient
 public class UserServiceApplication {
